ref: Apply codereview, refactor to project events base chart
